# fix: Correct VectorDBInitializer.search() method signature in vector backend